requirements.txt
autoattack==0.1
dassl.egg==info
einops==0.8.0
ftfy==6.3.0
numpy==1.24.3
Pillow==8.3.2
Pillow==11.0.0
regex==2023.10.3
scipy==1.14.1
setuptools==59.5.0
torch==1.10.1
torchvision==0.11.2
tqdm==4.66.1
yacs==0.1.8
